requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
tenacity==8.3.0
lxml==5.4.0
python-dotenv==1.2.2
requests==2.33.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
